requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
ninja==1.11.1.1
sqlparse==0.5.0
html-sanitizer==2.4.4